+++ src/runtime/index.ts
@@ -2,14 +2,28 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { CellEvent, NotebookContent } from "./types";
+import { CellEvent, NotebookContent, CellTypeDefinition } from "../types";
 import { ConsoleCatcher } from "../console/console";
 import { CellElement } from "../components/cell";
 import { StarboardNotebook } from "../components/notebook";
 
-export * from "./types";
+export * from "../types";
 
+/**
+ * Runtime is the main state/store for a notebook.
+ */
 export interface Runtime {
+
+    /**
+     * The state of the notebook that exactly describes the text in the notebook.
+     */
+    content: NotebookContent;
+
+
+    /**
+     * Map of registered cell types, indexed by cellType (e.g. "js").
+     */
+    cellTypes: Map<string, CellTypeDefinition>;
 
     /**
      * Contains HTML elements in this notebook runtime.
@@ -23,14 +37,13 @@ export interface Runtime {
      * Event bus for a notebook, used to propagate messages upwards such as "focus on the next cell".
      */
     emit: (e: CellEvent) => void;
-
-    /**
-     * The state of the notebook that exactly describes the text in the notebook.
-     */
-    content: NotebookContent;
-
     /**
      * Used to coordinate listening to the console hook.
      */
     consoleCatcher: ConsoleCatcher;
+
+    /**
+     * Version of Starboard Notebook
+     */
+    version: string;
 }
