+++ src/types.ts
@@ -2,8 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { Runtime } from ".";
-import { BaseCellHandler } from "../cellHandler/base";
+import { Runtime } from "./runtime";
 
 /**
  * Events that can be sent from the cell for central handling in the notebook component.
